restructure `scarb-completions` extension
